+++ app/src/main/java/de/ph1b/audiobook/features/bookPlaying/BookPlayPresenter.kt
@@ -85,7 +85,7 @@ class BookPlayPresenter(private val bookId: UUID) : BookPlayMvp.Presenter() {
     playerController.execute(PlayerCommand.Previous)
   }
 
-  override fun seekTo(position: Int, file: File?) {
+  override fun seekTo(position: Long, file: File?) {
     Timber.i("seekTo position$position, file$file")
     val book = bookRepository.bookById(bookId)
       ?: return
